feat(front-kit-3): extracted FooterSectionLayout into separate file
