+++ packages/falcon-ui-kit/src/Footer/FooterSectionListLayout.tsx
@@ -15,26 +15,3 @@ export const FooterSectionListLayout = themed({
     }
   }
 });
-
-export const FooterSectionLayout = themed({
-  tag: Box,
-  defaultTheme: {
-    footerSection: {
-      p: 'md',
-      display: 'flex',
-      flexDirection: 'column',
-      justifyContent: 'flex-start',
-      css: {
-        minWidth: 250,
-        textAlign: {
-          md: 'unset',
-          xs: 'center'
-        },
-        alignItems: {
-          md: 'unset',
-          xs: 'center'
-        }
-      }
-    }
-  }
-});
