+++ packages/falcon-ui-kit/src/Footer/index.ts
@@ -2,3 +2,4 @@ export * from './Footer';
 export * from './FooterBanner';
 export * from './Copyright';
 export * from './FooterSectionListLayout';
+export * from './FooterSectionLayout';
